+++ com.amitinside.featureflags.provider/src/main/java/com/amitinside/featureflags/provider/ServicePropertyActivationStrategy.java
@@ -7,9 +7,9 @@
  * http://www.eclipse.org/legal/epl-v10.html
  *
  *******************************************************************************/
-package com.amitinside.featureflags.internal;
+package com.amitinside.featureflags.provider;
 
-import static com.amitinside.featureflags.Constants.STRATEGY_SYSTEM_PROPERTY_PID;
+import static com.amitinside.featureflags.Constants.STRATEGY_SERVICE_PROPERTY_PID;
 import static org.osgi.service.component.annotations.ConfigurationPolicy.REQUIRE;
 
 import java.util.Map;
@@ -21,15 +21,16 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Modified;
 
 import com.amitinside.featureflags.Strategizable;
+import com.amitinside.featureflags.feature.Feature;
+import com.amitinside.featureflags.feature.group.FeatureGroup;
 import com.amitinside.featureflags.strategy.ActivationStrategy;
-import com.google.common.collect.Maps;
 
 /**
  * This strategy is responsible for checking configured property key and value in the
- * system configured properties.
+ * {@link Feature} or {@link FeatureGroup}'s OSGi service property.
  */
-@Component(name = "ConfiguredSystemPropertyStrategy", immediate = true, configurationPolicy = REQUIRE, configurationPid = STRATEGY_SYSTEM_PROPERTY_PID, service = ActivationStrategy.class)
-public final class SystemPropertyActivationStrategy extends AbstractPropertyActivationStrategy {
+@Component(name = "ConfiguredServicePropertyStrategy", immediate = true, configurationPolicy = REQUIRE, configurationPid = STRATEGY_SERVICE_PROPERTY_PID, service = ActivationStrategy.class)
+public final class ServicePropertyActivationStrategy extends AbstractPropertyActivationStrategy {
 
     @Override
     @Activate
@@ -51,10 +52,9 @@ public final class SystemPropertyActivationStrategy extends AbstractPropertyActi
         if (key == null || value == null) {
             return false;
         }
-        final Map<String, String> props = Maps.fromProperties(System.getProperties());
-        for (final Entry<String, String> entry : props.entrySet()) {
+        for (final Entry<String, Object> entry : properties.entrySet()) {
             final String k = entry.getKey();
-            final String v = entry.getValue();
+            final String v = String.valueOf(entry.getValue());
             if (Pattern.matches(key, k) && Pattern.matches(value, v)) {
                 return true;
             }
