+++ com.amitinside.featureflags.provider/src/test/java/com/amitinside/featureflags/provider/ConfigTest.java
@@ -7,7 +7,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  *
  *******************************************************************************/
-package com.amitinside.featureflags.internal;
+package com.amitinside.featureflags.provider;
 
 import static com.amitinside.featureflags.Config.*;
 import static org.junit.Assert.*;
