+++ com.amitinside.featureflags.provider/src/test/java/com/amitinside/featureflags/provider/FactoryTest.java
@@ -7,7 +7,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  *
  *******************************************************************************/
-package com.amitinside.featureflags.internal;
+package com.amitinside.featureflags.provider;
 
 import org.junit.Test;
